+++ packages/cdk8s-cli/lib/import/base.ts
@@ -1,8 +1,8 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import { CodeMaker } from 'codemaker';
-import { withTempDir, shell } from '../util';
-import { jsiiCompile } from './jsii';
+import { mkdtemp } from '../util';
+import * as srcmak from 'jsii-srcmak';
 
 export enum Language {
   TYPESCRIPT = 'typescript',
@@ -17,10 +17,17 @@ export interface ImportOptions {
   readonly moduleNamePrefix?: string;
   readonly targetLanguage: Language;
   readonly outdir: string;
+
+  /**
+   * Path to copy the output .jsii file.
+   * @default - jsii file is not emitted
+   */
+  readonly outputJsii?: string;
 }
 
 export abstract class ImportBase {
   public abstract get moduleNames(): string[];
+
   protected abstract async generateTypeScript(code: CodeMaker, moduleName?: string): Promise<void>;
 
   public async import(options: ImportOptions) {
@@ -31,7 +38,14 @@ export abstract class ImportBase {
     const isTypescript = options.targetLanguage === Language.TYPESCRIPT
     const { moduleNamePrefix } = options;
 
+    if (this.moduleNames.length === 0) {
+      console.error(`warning: no definitions to import`);
+    }
+
     for (const name of this.moduleNames) {
+      // output the name of the imported resource
+      console.error(name);
+
       const fileName = moduleNamePrefix ? `${moduleNamePrefix}-${name}.ts` : `${name}.ts`;
       code.openFile(fileName);
       code.indentation = 2;
@@ -41,45 +55,39 @@ export abstract class ImportBase {
       if (isTypescript) {
         await code.save(outdir);
       }
-    }
 
-    if (isTypescript) return;
+      if (!isTypescript || options.outputJsii) {
+        await mkdtemp(async staging => {
 
-    for (const name of this.moduleNames) {
-      // this is not typescript, so we generate in a staging directory and harvest the code
-      await withTempDir('importer', async () => {
-        await code.save('.');
-        await jsiiCompile('.', {
-          main: name,
-          moduleNamePrefix,
-          name,
+          // this is not typescript, so we generate in a staging directory and
+          // use jsii-srcmak to compile and extract the language-specific source
+          // into our project.
+          await code.save(staging);
+
+          // these are the module dependencies we compile against
+          const deps = [ '@types/node', 'constructs', 'cdk8s' ];
+
+          const opts: srcmak.Options = {
+            entrypoint: fileName,
+            deps: deps.map(dep => path.dirname(require.resolve(`${dep}/package.json`))),
+          };
+
+          // used for testing.
+          if (options.outputJsii) {
+            opts.jsii = { path: options.outputJsii };
+          }
+
+          // python!
+          if (options.targetLanguage === Language.PYTHON) {
+            opts.python = {
+              outdir: outdir,
+              moduleName: moduleNamePrefix ? `${moduleNamePrefix}.${name}` : name
+            };
+          }
+
+          await srcmak.srcmak(staging, opts);
         });
-
-        const pacmak = require.resolve('jsii-pacmak/bin/jsii-pacmak');
-        await shell(pacmak, [ '--target', options.targetLanguage, '--code-only' ]);
-        await this.harvestCode(options, outdir, name);
-      });
+      }
     }
-  }
-
-  private async harvestCode(options: ImportOptions, targetdir: string, moduleName: string) {
-    const { moduleNamePrefix } = options
-    switch (options.targetLanguage) {
-      case Language.TYPESCRIPT:
-        throw new Error('no op for typescript');
-  
-      case Language.PYTHON:
-        await this.harvestPython(targetdir, moduleNamePrefix ? moduleNamePrefix : moduleName);
-        break;
-  
-      default:
-        throw new Error(`unsupported language ${options.targetLanguage} (yet)`);
-    }
-  
-  }
-
-  private async harvestPython(targetdir: string, moduleName: string) {
-    const target = path.join(targetdir, moduleName);
-    await fs.move(`dist/python/src/${moduleName}`, target, { overwrite: true });
   }
 }