+++ packages/cdk8s-cli/lib/import/crd.ts
@@ -1,17 +1,16 @@
 import { TypeGenerator } from './type-generator';
 import { ImportBase } from './base';
 import { CodeMaker } from 'codemaker';
-import { httpGet, httpsGet } from '../util';
+import { download } from '../util';
 import * as yaml from 'yaml';
-import * as fs from 'fs-extra';
-import * as path from 'path';
 import { ImportSpec } from '../config';
 
 const CRD_KIND = 'CustomResourceDefinition';
 
-export interface CustomResourceApiObject {
+export interface ManifestObjectDefinition {
   apiVersion?: string;
   kind?: string;
+  items?: ManifestObjectDefinition[]; // if `kind` is "List"
   metadata?: {
     name?: string;
   };
@@ -19,10 +18,16 @@ export interface CustomResourceApiObject {
     group: string;
     names: {
       kind: string;
+      [key: string]: any;
     };
-    versions?: Array<{ name: string; schema?: { openAPIV3Schema?: any } }>;
+    versions?: Array<{
+      name: string;
+      schema?: { openAPIV3Schema?: any };
+      [key: string]: any;
+    }>;
     version?: string;
     validation?: { openAPIV3Schema?: any };
+    [key: string]: any;
   };
 }
 
@@ -39,7 +44,7 @@ export class CustomResourceDefinition {
   private readonly kind: string;
   private readonly fqn: string;
 
-  constructor(manifest: CustomResourceApiObject) {
+  constructor(manifest: ManifestObjectDefinition) {
     const apiVersion = manifest?.apiVersion ?? 'undefined';
     assert(SUPPORTED_API_VERSIONS.includes(apiVersion), `"apiVersion" is "${apiVersion}" but it should be one of: ${SUPPORTED_API_VERSIONS.map(x => `"${x}"`).join(', ')}`);
     assert(manifest.kind === CRD_KIND, `"kind" must be "${CRD_KIND}"`);
@@ -54,10 +59,6 @@ export class CustomResourceDefinition {
       throw new Error(`unable to determine CRD version`);
     }
 
-    if (!manifest.metadata?.name) {
-      throw new Error(`"metadata.name" is required`);
-    }
-
     const schema = typeof version === 'string'
       ? spec.validation?.openAPIV3Schema
       : version?.schema?.openAPIV3Schema ?? spec.validation?.openAPIV3Schema;
@@ -70,7 +71,7 @@ export class CustomResourceDefinition {
   }
 
   public get moduleName() {
-    return this.kind.toLocaleLowerCase();
+    return `${this.group}/${this.kind.toLocaleLowerCase()}`;
   }
 
   public async generateTypeScript(code: CodeMaker) {
@@ -93,37 +94,50 @@ export class CustomResourceDefinition {
 }
 
 export class ImportCustomResourceDefinition extends ImportBase {
-  public static async match(importSpec: ImportSpec): Promise<undefined | CustomResourceApiObject[]> {
+  public static async match(importSpec: ImportSpec): Promise<undefined | ManifestObjectDefinition[]> {
     const { source } = importSpec;
-    let manifest;
-    if (source.startsWith('https://')) {
-      manifest = await httpsGet(source);
-    } else if (source.startsWith('http://')) {
-        manifest = await httpGet(source)
-    } else if (path.extname(source) === '.yaml' || path.extname(source) === '.yml' || path.extname(source) === '.json') {
-      if (!(await fs.pathExists(source))) {
-        throw new Error(`can't find file ${source}`);
-      }
-
-      manifest = await fs.readFile(source, 'utf-8');
-    }
-
-    if (!manifest) {
-      return undefined;
-    }
-
+    const manifest = await download(source);
     return yaml.parseAllDocuments(manifest).map((doc: yaml.ast.Document) => doc.toJSON());
   }
 
   private readonly defs: CustomResourceDefinition[] = [];
 
-  constructor(manifest: CustomResourceApiObject[]) {
+  constructor(manifest: ManifestObjectDefinition[]) {
     super();
 
-    this.defs = manifest
-      .filter(obj => obj) // filter empty docs in the manifest
-      .filter(obj => obj.kind === CRD_KIND) // filter non-CRD resources from the manifest
-      .map(obj => new CustomResourceDefinition(obj));
+    const defs: Record<string, CustomResourceDefinition> = { };
+
+    const extractCRDs = (objects: ManifestObjectDefinition[] = []) => {
+      for (const obj of objects) {
+        // filter empty docs in the manifest
+       if (!obj) {
+         continue;
+       }
+
+       // found a crd, yey!
+       if (obj.kind === CRD_KIND) {
+          const crd = new CustomResourceDefinition(obj);
+          const key = crd.moduleName;
+
+          if (key in defs) {
+            throw new Error(`${key} already exists`);
+          }
+
+          defs[key] = crd;
+          continue;
+       }
+
+       // recurse into lists
+       if (obj.kind === 'List') {
+         extractCRDs(obj.items);
+         continue;
+       }
+     }
+    };
+
+    extractCRDs(manifest);
+
+    this.defs = Object.values(defs);
   }
 
   public get moduleNames() {