+++ packages/cdk8s-cli/lib/util.ts
@@ -1,38 +1,61 @@
 import * as http from 'http';
 import * as https from 'https';
 import { spawn, SpawnOptions } from 'child_process';
+import { parse } from 'url';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
 
 export async function shell(program: string, args: string[] = [], options: SpawnOptions = { }) {
+  const command = `"${program} ${args.join(' ')}" at ${path.resolve(options.cwd ?? '.')}`;
   return new Promise((ok, ko) => {
     const child = spawn(program, args, { stdio: 'inherit', ...options });
-    child.once('error', ko);
+    child.once('error', err => {
+      throw new Error(`command ${command} failed: ${err}`);
+    });
     child.once('exit', code => {
-      if (code === 0) { return ok(); }
-      else { return ko(new Error(`non-zero exit code ${code}`)); }
+      if (code === 0) {
+        return ok();
+      }
+      else {
+        return ko(new Error(`command ${command} returned a non-zero exit code ${code}`));
+      }
     });
   });
 }
 
-export async function withTempDir(dirname: string, closure: () => Promise<void>) {
-  const prevdir = process.cwd();
-  const parent = await fs.mkdtemp(path.join(os.tmpdir(), 'cdk8s.'));
-  const workdir = path.join(parent, dirname);
-  await fs.mkdirp(workdir);
+export async function mkdtemp(closure: (dir: string) => Promise<void>) {
+  const workdir = await fs.mkdtemp(path.join(os.tmpdir(), 'cdk8s-'));
   try {
-    process.chdir(workdir);
-    await closure();
+    await closure(workdir);
   } finally {
-    process.chdir(prevdir);
-    await fs.remove(parent);
+    await fs.remove(workdir);
   }
 }
 
-async function get(url: string, protocol: typeof http | typeof https = https): Promise<string> {
+export async function download(url: string): Promise<string> {
+  let client: typeof http | typeof https;
+  const proto = parse(url).protocol;
+
+  if (!proto || proto === 'file:') {
+    return await fs.readFile(url, 'utf-8');
+  }
+
+  switch (proto) {
+    case 'https:':
+      client = https;
+      break;
+
+    case 'http:':
+      client = http;
+      break;
+
+    default:
+      throw new Error(`unsupported protocol ${proto}`);
+  }
+
   return new Promise((ok, ko) => {
-    const req = protocol.get(url, async res => {
+    const req = client.get(url, res => {
       switch(res.statusCode) {
         case 200: {
           const data = new Array<Buffer>();
@@ -45,7 +68,7 @@ async function get(url: string, protocol: typeof http | typeof https = https): P
         case 301:
         case 302: {
           if (res.headers.location) {
-            await ok(get(res.headers.location, protocol));
+            ok(download(res.headers.location));
           }
           break;
         }
@@ -59,12 +82,4 @@ async function get(url: string, protocol: typeof http | typeof https = https): P
     req.once('error', ko);
     req.end();
   });
-}
-
-export async function httpGet(url: string): Promise<string> {
-  return get(url, http)
-}
-
-export async function httpsGet(url: string): Promise<string> {
-  return get(url)
 }