+++ packages/cdk8s-cli/test/import/import-crd.test.ts
@@ -10,8 +10,8 @@ expectImportMatchSnapshotCustomResource('mixed_crd.yaml');
 expectImportMatchSnapshotCustomResource('prometheus.yaml');
 
 test('fails if CRDs api version is not supported', () => {
-  expect(() => new ImportCustomResourceDefinition([{ 
-    apiVersion: 'voo' ,
+  expect(() => new ImportCustomResourceDefinition([{
+    apiVersion: 'voo',
     kind: 'CustomResourceDefinition'
   }])).toThrow('invalid CustomResourceDefinition manifest: "apiVersion" is "voo" but it should be one of: "apiextensions.k8s.io/v1beta1", "apiextensions.k8s.io/v1"');
 });
@@ -25,7 +25,7 @@ test('fails if manifest does not have a "spec" field', () => {
 
 test('fails if one apiObject in multiObject CRD is not a valid CRD', async () => {
   expect(() => new ImportCustomResourceDefinition([
-    { 
+    {
       apiVersion: 'apiextensions.k8s.io/v1beta1',
       kind: 'CustomResourceDefinition',
       metadata: {
@@ -51,6 +51,151 @@ test('fails if one apiObject in multiObject CRD is not a valid CRD', async () =>
       kind: 'CustomResourceDefinition'
     }
   ])).toThrow('manifest does not have a "spec" attribute');
+});
+
+test('can import a "List" of CRDs (kubectl get crds -o json)', () => {
+  const importer = new ImportCustomResourceDefinition([
+    {
+      kind: 'List',
+      items: [
+        {
+          "apiVersion": "apiextensions.k8s.io/v1beta1",
+          "kind": "CustomResourceDefinition",
+          "metadata": {
+            "name": "crontabs.stable.example.com"
+          },
+          "spec": {
+            "group": "stable.example.com",
+            "versions": [
+              {
+                "name": "v1",
+                "served": true,
+                "storage": true
+              }
+            ],
+            "scope": "Namespaced",
+            "names": {
+              "plural": "crontabs",
+              "singular": "crontab",
+              "kind": "OtherCronTab",
+              "shortNames": [
+                "ct"
+              ]
+            },
+            "preserveUnknownFields": false,
+            "validation": {
+              "openAPIV3Schema": {
+                "type": "object",
+                "properties": {
+                  "spec": {
+                    "type": "object",
+                    "properties": {
+                      "cronSpec": {
+                        "type": "string"
+                      },
+                      "image": {
+                        "type": "string"
+                      },
+                      "replicas": {
+                        "type": "integer"
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        },
+        {}, // verify that we skip empty
+        {
+          "kind": "List",
+          "items": []
+        },
+
+        // nested lists
+        {
+          "kind": "List",
+          "items": [
+            {
+              "apiVersion": "apiextensions.k8s.io/v1beta1",
+              "kind": "CustomResourceDefinition",
+              "spec": {
+                "group": "foo.bar",
+                "version": "v1",
+                "names": {
+                  "kind": "foo"
+                }
+              }
+            },
+
+            // skip non-CRD
+            {
+              "apiVersion": "apiextensions.k8s.io/v1beta1",
+              "kind": "NonCustomResourceDefinition",
+              "spec": {
+                "group": "foo.bar",
+                "names": { "kind": "foo" },
+              }
+            }
+          ]
+        },
+        {
+          "apiVersion": "apiextensions.k8s.io/v1beta1",
+          "kind": "CustomResourceDefinition",
+          "metadata": {
+            "name": "crontabs.stable.example.com"
+          },
+          "spec": {
+            "group": "stable.example.com",
+            "versions": [
+              {
+                "name": "v1",
+                "served": true,
+                "storage": true
+              }
+            ],
+            "scope": "Namespaced",
+            "names": {
+              "plural": "crontabs",
+              "singular": "crontab",
+              "kind": "CronTab",
+              "shortNames": [
+                "ct"
+              ]
+            },
+            "preserveUnknownFields": false,
+            "validation": {
+              "openAPIV3Schema": {
+                "type": "object",
+                "properties": {
+                  "spec": {
+                    "type": "object",
+                    "properties": {
+                      "cronSpec": {
+                        "type": "string"
+                      },
+                      "image": {
+                        "type": "string"
+                      },
+                      "replicas": {
+                        "type": "integer"
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      ]
+    }
+  ]);
+
+  expect(importer.moduleNames).toEqual([
+    'stable.example.com/othercrontab',
+    'foo.bar/foo',
+    'stable.example.com/crontab'
+  ]);
 });
 
 function loadFixture(fileName: string) {