+++ packages/cdk8s/lib/index.ts
@@ -2,3 +2,5 @@ export * from './api-object';
 export * from './chart';
 export * from './testing';
 export * from './app';
+export * from './include';
+export * from './yaml';